#11 - Rename CarouselItem to WeatherTimelineItem
